fix: widen message ids to int64 / BIGSERIAL
messages.id was SERIAL (int32), which silently corrupts message identity
once the sequence wraps — months at sustained high publish rates.
The Go client mirrored the schema's int and would have overflowed too;
the TS client received BIGINT as string by default in node-pg, an easy
footgun.

SQL: messages.id → BIGSERIAL; queue_messages.message_id and
dead_letter_queue.message_id → BIGINT. All function signatures
referencing a message id widen accordingly (publish/consume returns,
ack/nack/release/set_vt parameters, set_vt_batch arrays, list_*/
get_message return shapes).

Go: int → int64 wherever a message id flows — Message.ID,
Publish*/PublishWithTx returns, MessageExtension.ID, MessageLock.ID,
QueueMessage/PublishedMessage/DLQMessage.MessageID, ack/nack/release/
setVt parameters, vtInfo.messageID, the internal heap map key, and
the SetVTBatch ids slice.

TS: register a process-global node-pg type parser for OID 20 (BIGINT)
so values come back as number — JS number covers up to 2^53, well
past the int32 limit we just escaped, so the existing
Message.id: number surface stays correct.

Round-trip tests in Go (TestMessageID_BeyondInt32) and TS (in
consumer.test.ts) bump messages_id_seq past 2^31 and verify
publish/consume preserves the id.

+++ postgremq-go/consumer_test.go
@@ -47,7 +47,7 @@ func TestConsumerStartAndReceive(t *testing.T) {
 		payload := []byte(fmt.Sprintf("{\"message\": \"message-%d\"}", i))
 		messageID, err := conn.Publish(ctx, topicName, payload)
 		require.NoError(t, err, "Publish failed")
-		assert.Greater(t, messageID, 0, "Expected valid message ID")
+		assert.Greater(t, messageID, int64(0), "Expected valid message ID")
 	}
 
 	// conn.Consume automatically starts the consumer.
@@ -150,7 +150,7 @@ func TestConsumerMultipleDistribution(t *testing.T) {
 	for i := 0; i < 10; i++ {
 		messageID, err := conn.Publish(ctx, topicName, []byte(fmt.Sprintf("\"msg-%d\"", i)))
 		require.NoError(t, err, "Publish failed")
-		assert.Greater(t, messageID, 0, "Expected valid message ID")
+		assert.Greater(t, messageID, int64(0), "Expected valid message ID")
 	}
 
 	consumer1, err := conn.Consume(ctx, queueName, postgremq.WithBatchSize(3))
@@ -161,7 +161,7 @@ func TestConsumerMultipleDistribution(t *testing.T) {
 	require.NoError(t, err, "Consume failed for consumer2")
 	defer consumer2.Stop()
 
-	receivedMsgs := make(map[int]bool)
+	receivedMsgs := make(map[int64]bool)
 	var mu sync.Mutex
 	wg := sync.WaitGroup{}
 	wg.Add(2)
@@ -272,14 +272,14 @@ func TestConsumerBufferedMessagesReleased(t *testing.T) {
 		payload := []byte(fmt.Sprintf("{\"msg\": \"msg-%d\"}", i))
 		messageID, err := conn.Publish(ctx, topicName, payload)
 		require.NoError(t, err, "Publish failed")
-		require.Greater(t, messageID, 0, "Expected valid message ID")
+		require.Greater(t, messageID, int64(0), "Expected valid message ID")
 	}
 
 	// Create a consumer with a BatchSize of 5.
 	consumer, err := conn.Consume(ctx, queueName, postgremq.WithBatchSize(5), postgremq.WithVT(10))
 	require.NoError(t, err, "Consume failed")
 
-	nackedMessages := make(map[int]struct{})
+	nackedMessages := make(map[int64]struct{})
 	// Read only 3 messages from the consumer and nack them.
 	for i := 0; i < 3; i++ {
 		msg := <-consumer.Messages()
